refactor: restructure site with sidebar layout and extract components
- Add left sidebar navigation with News (default) and About Me tabs
- Move original single-page content into pages/AboutMe.svelte
- Add pages/News.svelte as a terminal-style placeholder
- Extract 6 components from the 1,583-line AboutMe monolith:
  TerminalBoot, GameOfLife, MotivationSection, HangingScroll,
  CreedPoster, QuotesTerminal
- Move motivationQuotes and bootSequence arrays to src/data/
- Fix 10 a11y violations: replace clickable <div> elements with
  <button> (7 image cards, 3 terminal dots); fix lightbox img event
