+++ src/components/Sidebar.svelte
@@ -1,0 +1,201 @@
+<script>
+  export let activeTab = 'news';
+
+  let mobileOpen = false;
+
+  const tabs = [
+    { id: 'news',    label: 'News',     icon: '📰' },
+    { id: 'aboutme', label: 'About Me', icon: '👤' },
+  ];
+
+  function selectTab(id) {
+    activeTab = id;
+    mobileOpen = false;
+  }
+
+  function toggleMobile() {
+    mobileOpen = !mobileOpen;
+  }
+</script>
+
+<style>
+  /* ── Hamburger (mobile only) ── */
+  .hamburger {
+    display: none;
+    position: fixed;
+    top: 14px;
+    left: 14px;
+    z-index: 300;
+    background: rgba(0, 5, 20, 0.95);
+    border: 1px solid rgba(0, 245, 255, 0.4);
+    color: #00f5ff;
+    font-size: 1.2rem;
+    line-height: 1;
+    padding: 0.4rem 0.6rem;
+    cursor: pointer;
+    border-radius: 2px;
+    transition: all 0.2s;
+  }
+
+  .hamburger:hover {
+    background: rgba(0, 245, 255, 0.1);
+    box-shadow: 0 0 12px rgba(0, 245, 255, 0.3);
+  }
+
+  /* ── Sidebar ── */
+  .sidebar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 220px;
+    height: 100vh;
+    background: rgba(0, 5, 20, 0.98);
+    border-right: 1px solid rgba(0, 245, 255, 0.25);
+    display: flex;
+    flex-direction: column;
+    z-index: 200;
+    box-shadow: 4px 0 24px rgba(0, 245, 255, 0.06);
+    transition: transform 0.28s cubic-bezier(0.4, 0, 0.2, 1);
+  }
+
+  /* ── Logo / Title ── */
+  .sidebar-logo {
+    padding: 1.8rem 1.5rem 1.4rem;
+    border-bottom: 1px solid rgba(0, 245, 255, 0.12);
+  }
+
+  .logo-text {
+    font-family: 'JetBrains Mono', 'SF Mono', monospace;
+    font-size: 1.3rem;
+    font-weight: 700;
+    color: #00f5ff;
+    letter-spacing: 4px;
+    text-transform: lowercase;
+    text-shadow: 0 0 16px rgba(0, 245, 255, 0.6);
+    display: block;
+  }
+
+  .logo-sub {
+    font-family: 'JetBrains Mono', monospace;
+    font-size: 0.65rem;
+    color: #8892b0;
+    letter-spacing: 2px;
+    margin-top: 0.25rem;
+    display: block;
+  }
+
+  /* ── Nav ── */
+  nav {
+    flex: 1;
+    padding: 1.2rem 0;
+    overflow-y: auto;
+  }
+
+  .nav-item {
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    padding: 0.75rem 1.5rem;
+    cursor: pointer;
+    color: #8892b0;
+    font-size: 0.9rem;
+    font-weight: 400;
+    letter-spacing: 0.5px;
+    position: relative;
+    transition: color 0.2s ease, background 0.2s ease;
+    border: none;
+    background: none;
+    width: 100%;
+    text-align: left;
+    border-left: 3px solid transparent;
+  }
+
+  .nav-item:hover {
+    color: #c0c8e8;
+    background: rgba(0, 245, 255, 0.05);
+  }
+
+  .nav-item.active {
+    color: #00f5ff;
+    background: rgba(0, 245, 255, 0.07);
+    border-left-color: #00f5ff;
+    text-shadow: 0 0 8px rgba(0, 245, 255, 0.4);
+  }
+
+  .nav-icon {
+    font-size: 1rem;
+    flex-shrink: 0;
+    width: 20px;
+    text-align: center;
+  }
+
+  /* ── Footer strip ── */
+  .sidebar-footer {
+    padding: 1rem 1.5rem;
+    border-top: 1px solid rgba(0, 245, 255, 0.1);
+    font-family: 'JetBrains Mono', monospace;
+    font-size: 0.6rem;
+    color: #404868;
+    letter-spacing: 1px;
+  }
+
+  /* ── Mobile overlay backdrop ── */
+  .backdrop {
+    display: none;
+    position: fixed;
+    inset: 0;
+    background: rgba(0, 0, 0, 0.7);
+    z-index: 199;
+  }
+
+  /* ── Responsive ── */
+  @media (max-width: 768px) {
+    .hamburger {
+      display: block;
+    }
+
+    .sidebar {
+      transform: translateX(-100%);
+    }
+
+    .sidebar.open {
+      transform: translateX(0);
+    }
+
+    .backdrop {
+      display: block;
+    }
+  }
+</style>
+
+<!-- Mobile hamburger -->
+<button class="hamburger" on:click={toggleMobile} aria-label="Toggle menu">
+  {mobileOpen ? '✕' : '☰'}
+</button>
+
+<!-- Backdrop (mobile) -->
+{#if mobileOpen}
+  <div class="backdrop" on:click={toggleMobile} role="presentation"></div>
+{/if}
+
+<aside class="sidebar" class:open={mobileOpen}>
+  <div class="sidebar-logo">
+    <span class="logo-text">qitpy</span>
+    <span class="logo-sub">// personal space</span>
+  </div>
+
+  <nav>
+    {#each tabs as tab (tab.id)}
+      <button
+        class="nav-item"
+        class:active={activeTab === tab.id}
+        on:click={() => selectTab(tab.id)}
+      >
+        <span class="nav-icon">{tab.icon}</span>
+        {tab.label}
+      </button>
+    {/each}
+  </nav>
+
+  <div class="sidebar-footer">v0.1.0 · qitpy</div>
+</aside>
