+++ src/data/boot.js
@@ -1,0 +1,7 @@
+export const bootSequence = [
+  '[hIE] Initializing...',
+  '[hIE] Behavioral model loaded...',
+  '[hIE] Owner registration: confirmed',
+  'ACCESS GRANTED — Welcome, visitor.',
+  '> _ <',
+];
