+++ src/data/quotes.js
@@ -1,0 +1,12 @@
+export const motivationQuotes = [
+  { text: "Discipline is doing what you hate to do, but doing it like you love it.", author: "Mike Tyson" },
+  { text: "Everyone has a plan until they get punched in the mouth.", author: "Mike Tyson" },
+  { text: "I'm not the same person I was when I bit that guy's ear off.", author: "Mike Tyson" },
+  { text: "The pain you feel today will be the strength you feel tomorrow.", author: "Unknown" },
+  { text: "Your only limit is your mind.", author: "Unknown" },
+  { text: "Be brutal with your effort, patient with your results.", author: "Unknown" },
+  { text: "Champions aren't made in gyms. Champions are made from something deep inside.", author: "Muhammad Ali" },
+  { text: "Pain is temporary. Quitting lasts forever.", author: "Lance Armstrong" },
+  { text: "Any sufficiently advanced technology is indistinguishable from magic.", author: "Arthur C. Clarke" },
+  { text: "Forget the mountain, just upgrade the gear. Fascination leads faster than any plan.", author: "makeitbetter" },
+];
